feat: #60 Connect BusStopView and ScannedJourneyInfoView
